Offset BC plot types
From what I can tell, BC symbols are offset when plotted on the graph. I
added an offset to each of them in what I think is the correct
direction. The offset is equal to the size of the symbol, meaning each
symbol rests against the line it's plotted on instead of being centered
on that line.