WAMP Zip - has Database, PHP , Stuff
My WAMP installation containing all its files. Will upload the
individual files later. Most of the important stuff is in the WWW
folder. It has a database with some sample patient info, and a page
which you can enter a name into and it will update the page with the
patient's info. Just need to merge it with our current index.html