Fixed duplication bug in audiogram graphing
I just re-wrote the logic that deletes symbols when they're supposed to
be deleted (and re-wrote the logic that determines when they're supposed
to be deleted). I changed a lot about the data structure of our points.
It works now. I found another, very small, bug that I'll fix sometime.