Fixed file indentation and fixed small audiogram bug
In the course of making the last fix I found a small bug where the graph
has frequencies for places it shouldn't. It's not totally fixed, but
it's no longer a problem. Now trying to plot in the big spaces between
lines towards the left of the graph will just do nothing.

That might be frustrating for users. Dunno. I'm sure at some point
someone will want to go through and fix it, but that's a problem for
later, if ever.